Move iCloud settings for Contacts from main settings to contacts settings page
