+++ UnstoppableWallet/UnstoppableWallet/Modules/ContactBook/ContactBookSettings/ContactBookSettingsModule.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+class ContactBookSettingsModule {
+
+    static var viewController: UIViewController? {
+        guard let manager = App.shared.contactManager else {
+            return nil
+        }
+        let service = ContactBookSettingsService(contactManager: manager)
+        let viewModel = ContactBookSettingsViewModel(service: service)
+        return ContactBookSettingsViewController(viewModel: viewModel)
+    }
+
+}
